# GODRIVER-2536 Return early in replaceErrors.