GODRIVER-2536

Returns `nil` in `replaceErrors` when `err == nil` to avoid costly reflection logic below. 

[Here's a benchmark](https://gist.github.com/benjirewis/8b088d422788a4beceeb5076652db256) I used to measure the difference in operations per second. The output from `benchstat` was:

```
name             old time/op    new time/op    delta
ReplaceErrors-8    12.2ns ± 1%     2.7ns ± 0%  -78.19%  (p=0.008 n=5+5)
```